backend/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
djangorestframework>=3.15.0
drf-spectacular>=0.27.2
django-cors-headers>=4.4.0
djangorestframework-simplejwt>=5.3.1